requirements.txt
streamlit
pymupdf
google-generativeai
python-dotenv
fpdf2
sqlalchemy
psycopg2-binary
